requirements.txt
streamlit==1.32.0
google-generativeai==0.8.6
python-dotenv==1.0.1
pypdf==4.0.0
python-docx==0.8.11
Pillow==10.1.0
pytesseract==0.3.10